Creating HTTP Client - Need direction on the process
